requirements.txt
Flask==2.2.2
torch==2.0.0 -f https://download.pytorch.org/whl/cpu/torch_stable.html
numpy==1.24.3
transformers==4.28.1
fuzzywuzzy==0.18.0
tqdm==4.65.0
python-Levenshtein==0.21.0
flask_httpauth==4.8.0
load-dotenv==0.1.0
sentence_transformers==2.2.2
datasets==2.12.0
